feat: 补全题目新增逻辑

+++ dk-club-subject/dk-club-domain/src/main/java/com/dk/subject/domain/bo/SubjectAnswerBO.java
@@ -19,7 +19,7 @@ public class SubjectAnswerBO implements Serializable {
     /**
      * 选项类型
      */
-    private Long optionType;
+    private Integer optionType;
     /**
      * 选项内容
      */
