+++ dk-club-subject/dk-club-infra/src/main/java/com/dk/subject/infra/basic/entity/SubjectBrief.java
@@ -26,7 +26,7 @@ public class SubjectBrief implements Serializable {
     /**
      * 题目id
      */
-    private Integer subjectId;
+    private Long subjectId;
     /**
      * 题目答案
      */
